+++ client/components/AuthorizationForms/LoginForm.jsx
@@ -14,8 +14,6 @@ import Stack from "@mui/material/Stack";
 import LoginIcon from '@mui/icons-material/Login';
 import { useMediaQuery, useTheme } from '@mui/material';
 
-import LoadingMessage from "../ErrorMessages/LoadingMessage"
-
 import { useLoginMutation, useGetAllUsersValidationQuery } from "../../redux/api";
 
 const LoginForm = () => {
