+++ client/components/Dashboards/AdminDashboard/MapPostsAdmin.jsx
@@ -16,10 +16,8 @@ import { Link } from "react-router-dom";
 
 import { useState } from "react";
 
-import { useGetReviewsQuery } from "../../redux/api";
-import { useDeleteReviewForUserMutation } from "../../redux/api";
-import LoadingMessage from "../ErrorMessages/LoadingMessage";
-// import { useDeleteReviewForUserMutation } from "../../redux/api";
+import { useGetReviewsQuery } from "../../../redux/api";
+import { useDeleteReviewForUserMutation } from "../../../redux/api";
 
 //<-----------------DELETE REVIEW HELPER FUNCTION------------------->
 
@@ -36,7 +34,7 @@ const MapPostsAdmin = () => {
         return <div> Oops, our own web equipment is broken. We should have the issue resolved soon! </div>
     }
     if (isLoading) {
-        return <div><LoadingMessage /></div>;
+        return <div> </div>;
     }
     if (error) {
         return <div>Error:{error.message}</div>;
