+++ client/components/Dashboards/UserDashboard/EditUser.jsx
@@ -9,9 +9,8 @@ import { motion } from "framer-motion";
 
 import { useNavigate, useParams } from "react-router-dom";
 
-import { usePatchUserMutation, useGetUserQuery } from "../../redux/api";
+import { usePatchUserMutation, useGetUserQuery } from "../../../redux/api";
 import { useState } from "react";
-import LoadingMessage from "../ErrorMessages/LoadingMessage";
 
 const EditUser = () => {
     const { id } = useParams();
@@ -41,7 +40,7 @@ const EditUser = () => {
         return <div>Error</div>
     }
     if (isLoading) {
-        return <div><LoadingMessage/></div>
+        return <div> </div>
     }
 
     return userData && (
