+++ client/components/EquipmentAndReviews/AllPosts.jsx
@@ -18,9 +18,8 @@ import { Link } from 'react-router-dom';
 import { useGetReviewsQuery } from '../../redux/api';
 
 import SearchBar from '../SearchEquipment/SearchBar';
-import LoadingMessage from '../ErrorMessages/LoadingMessage';
 import EquipmentName from './EquipmentName';
-import MapAllEquipment from '../EquipmentsWithReviews/MapAllEquipment';
+import MapAllEquipment from '../SingleEquipment/MapAllEquipment';
 
 const AllPosts = () => {
     const theme = useTheme();
@@ -29,7 +28,7 @@ const AllPosts = () => {
     const { data, error, isLoading } = useGetReviewsQuery()
 
     if (isLoading) {
-        return <div><LoadingMessage /></div>
+        return <div> </div>
     }
     if (error) {
         return <div> Sorry! There's a problem loading the reviews. </div>
