+++ client/components/EquipmentAndReviews/EquipmentName.jsx
