+++ client/components/EquipmentAndReviews/MapAllEquipment.jsx
@@ -12,7 +12,6 @@ import "react-multi-carousel/lib/styles.css";
 import { Link } from "react-router-dom";
 
 import { useGetEquipmentQuery } from "../../redux/api"
-import LoadingMessage from "../ErrorMessages/LoadingMessage";
 
 const MapAllEquipment = () => {
 
@@ -37,7 +36,7 @@ const MapAllEquipment = () => {
 
     const { data, error, isLoading } = useGetEquipmentQuery();
     if (isLoading) {
-        return <div><LoadingMessage /></div>
+        return <div> </div>
     }
     if (error) {
         return <div> Sorry! There's a problem loading the equipment. </div>
