+++ client/components/EquipmentAndReviews/SearchEquipment/SearchResults.jsx
