--- client/components/ErrorMessages/LoadingMessage.jsx
@@ -1,10 +1,0 @@
-import CircularProgress from '@mui/material/CircularProgress';
-
-const LoadingMessage = () => {
-    return(
-        <>
-        <CircularProgress />
-        </>
-    )
-}
-export default LoadingMessage
