+++ client/components/HomePage/images/cookingEquipmentLogo.png
