+++ client/components/HomePage/images/kitchenEquipmentBackground.png
